feat(playout): don't clear liquidsoap queue on liquidsoap startup (#3163)
### Description

When liquidsoap starts, it sends an API request to the legacy API which
triggers a rabbitmq message to the playout service to update the
liquidsoap queues and make sure they are correct. The removed line is
not necessary and will cause issues if there is more than one liquidsoap
service. It is not necessary because the playout service automatically
prunes the queue and will add/remove only what is necessary and does not
require the whole queue to be cleared. If the whole queue is cleared and
there are multiple liquid soap services, every LQ service will go silent
and then fade to normal volume in the middle of playout. Removing this
line has no logical effect as all the work it does is immediately redone
and removal fixes the re-cue-in issue when you have multiple liquidsoap
services all running playout.

So, this line in the best of cases is useless and does nothing, and in
the worst of cases causes playout issues, hence its removal.

**This is a new feature**:

No

**I have updated the documentation to reflect these changes**:

Bug fix

### Testing Notes

**What I did:**

I setup multiple playout/liquidsoap service pairs, and heard that when
one is restarted all of them go quiet and fade in from their current
play cursor. I then made the change and verified that only the restarted
service faded in.

**How you can replicate my testing:**

See above section

### **Links**

Related to #3161 and #3162